[feature/REW-001] model 분리
